nothing interesting
-Switched to UTF8
-Updated to do

This might be the last commit to this repository. This project was meant to teach me how servers work and it has. It still has a lot of things that I could have done better that I might come back to. But right now I feel like it is at a point at which I'm happy with. And thus I have decided to start working on the next project.